requirements.txt
Django==1.8
Scrapy==0.24.5
Twisted==15.1.0
argparse==1.2.1
cffi==0.9.2
cryptography==0.8.2
cssselect==0.9.1
dj-database-url==0.3.0
dj-static==0.0.6
django-extensions==1.5.2
django-phonenumber-field==0.7.2
django-registration-redux==1.1
django-toolbelt==0.0.1
djangorestframework==3.1.1
enum34==1.0.4
gunicorn==19.3.0
html2text==2015.4.14
lxml==3.4.2
phonenumbers==7.0.4
psycopg2==2.6
pyOpenSSL==0.15.1
pyasn1==0.1.7
pycparser==2.10
python-dateutil==2.4.2
python-mimeparse==0.1.4
queuelib==1.2.2
requests==2.6.0
six==1.9.0
static3==0.5.1
w3lib==1.11.0
whitenoise==1.0.6
wsgiref==0.1.2
zope.interface==4.1.2
